Closes #122, closes #125, closes #126, closes #127.

+++ common-diffusion/type.h
@@ -33,7 +33,7 @@
  choice propagates throughout the code, and may significantly affect runtime
  on GPU hardware.
 */
-typedef float fp_t;
+typedef double fp_t;
 
 /**
  Container for timing data
